Day19/requirements.txt
torch
torchvision
fastapi
uvicorn
pillow
numpy
python-multipart
opencv-python